requirements.txt
beautifulsoup4==4.8.1
dj-static==0.0.6
Django==2.2.8
django-appconf==1.0.3
django-autocomplete-light==3.5.0
django-bootstrap-datepicker-plus==3.0.5
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-js-asset==1.2.2
django-s3direct==1.1.5
django-select2==6.3.1
django-widget-tweaks==1.4.3
djangorestframework==3.10.3
djangorestframework-datatables==0.5.0
docutils==0.15.2
gevent==20.5.0
greenlet==0.4.15
gunicorn==19.9.0
Markdown==3.1.1
psycopg2==2.8.5
Pillow==6.2.1
python-dateutil==2.8.0
python-decouple==3.0
pytz==2019.3
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.7
whitenoise==3.2
